Moved linalg functions from multi.c -> linalg.c and made corresponding header file changes
