+++ c-backend/linalg.c
@@ -108,6 +108,191 @@ Vector multiply_matrix_vector(Matrix X, Vector y) {
     return z;
 }
 
+// res = x_T * y
+double multiply_vector_vector(Vector x, Vector y) {
+    double res = 0.0f;
+    int i;
+
+    if (x.size != y.size) {
+        printf("ERROR in dot product of 2 vectors. Dimensions of vector x is %dx1 and of vector y is %dx1\n", x.size, y.size);
+        return 0.0f;
+    }
+
+    for (i = 0; i < x.size; i++) {
+        res += x.data[i] * y.data[i];
+    }
+
+    return res;
+}
+
+// Return column i of matrix X
+Vector get_column(Matrix X, int i) {
+    Vector res;
+    int j;
+    res.size = X.n;
+    res.data = (double*)malloc(res.size * sizeof(double));
+    
+    for (j = 0; j < X.n; j++) {
+        res.data[j] = X.data[j * X.m + i];
+    }
+
+    return res;
+}
+
+// Return the magnitude of vector x
+double get_magnitude(Vector x) {
+    int i;
+    double mag = 0.0f;
+
+    for (i = 0; i < x.size; i++) {
+        mag += x.data[i] * x.data[i];
+    }
+    mag = pow(mag, 1.0 / 2.0);
+
+    return mag;
+}
+
+// I will only use these functionalities in place and so the following sometimes take pass by reference (pointer) fields
+
+// x = x - y
+void subtract_vector_vector_inplace(Vector *x, Vector y) {
+    int i;
+
+    if (x->size != y.size) {
+        printf("ERROR in subtracting 2 vectors. Dimensions of vector x is %dx1 and of vector y is %dx1\n", x->size, y.size);
+        return;
+    }
+
+    for (i = 0; i < x->size; i++) {
+        x->data[i] -= y.data[i];
+    }
+
+    return;
+}
+
+// x = r * x
+void multiply_scalar_vector_inplace(double scalar, Vector *x) {
+    int i;
+
+    for (i = 0; i < x->size; i++) {
+        x->data[i] *= scalar;
+    }
+
+    return;
+}
+
+// X_i = col where i = col_idx
+void copy_column_to_matrix_inplace(Vector col, Matrix *X, int col_idx) {
+    int i;
+
+    if (col.size != X->n) {
+        printf("ERROR in inserting column into matrix. Dimensions of vector is %dx1 and of destination matrix is %dx%d\n", col.size, X->n, X->m);
+        return;
+    }
+
+    for (i = 0; i < col.size; i++) {
+        X->data[i * X->m + col_idx] = col.data[i];
+    }
+
+    return;
+}
+
+// ADVANCED TECHNIQUES ------
+
+// Solve upper triangular system via back substitution: UT * x = y
+Vector solve_back_sub(Matrix UT, Vector y)  {
+    Vector x; int i,j;
+    x.size = UT.m;
+    x.data = (double*)malloc(sizeof(double)*x.size);
+
+    // Input validation
+    if (UT.n != UT.m) {
+        printf("ERROR in solving upper-triangular system UT*x = y. Dimensions of matrix UT is %dx%d, and it should be square for a consistent system.\n", UT.n, UT.m);
+        return x;
+    }
+
+    if (UT.n != y.size) {
+        printf("ERROR in solving upper-triangular system UT*x = y. Dimensions of matrix UT is %dx%d and of vector y is %dx1\n", UT.n, UT.m, y.size);
+        return x;
+    }
+
+    if (is_upper_triangular(&UT) == 0) {
+        printf("ERROR in solving upper-triangular system UT*x = y. Matrix UT is not upper triangular.\n");
+        return x;
+    }
+
+    // Back substitution
+    for (i = x.size - 1; i >= 0; i--) {
+        double coeff = UT.data[i * UT.m + i];
+        double res = y.data[i];
+
+        // Avoid division by 0 error
+        if (coeff == 0.0f) {
+            printf("BEWARE: in solving upper-triangular system UT*x = y. There exists a 0 on the diagonal of matrix UT, making the system have infinite solutions.\n");
+            continue;
+        }
+        
+        // Substitute discovered values
+        for (j = UT.m-1; j > i; j--) {
+            res -= UT.data[i * UT.m + j] * x.data[j];
+        }
+        res /= coeff;
+        x.data[i] = res;
+    }
+
+    return x;
+}
+
+// QR factorisation via Classical Gram-Schmidt
+QR QR_factorise(Matrix X) {
+    QR res;
+    double magnitude;
+
+    res.Q.n = X.n;
+    res.Q.m = X.m;
+    res.R.n = X.m;
+    res.R.m = X.m;
+    res.Q.data = (double*)malloc(res.Q.n*res.Q.m*sizeof(double));
+    res.R.data = (double*)malloc(res.R.n*res.R.m*sizeof(double));
+
+    // printf("%d x %d, %d x %d", res.Q.n, res.Q.m, res.R.n, res.R.m);
+
+    int i, j;
+
+    // loop over the columns of X
+    for (i = 0; i < X.m; i++) {
+        // generate corresponding orthonormal column of Q and necessary entries in R
+        Vector X_i = get_column(X, i);
+        Vector Q_i = get_column(X, i); // Q_i = X_i
+
+        for (j = 0; j < i; j++) {
+            Vector Q_j = get_column(res.Q, j);
+            // r_ji = Q_j • X_i
+            double r_ji = multiply_vector_vector(Q_j, X_i);
+            // save r_ji to matrix R
+            res.R.data[j*res.R.m + i] = r_ji;
+
+            // Q_i = Q_i - r_ji * Q_j
+            multiply_scalar_vector_inplace(r_ji, &Q_j);
+            subtract_vector_vector_inplace(&Q_i, Q_j);
+        }
+
+        // Q_i = Q_i / |Q_i|
+        // r_ii = |Q_i|
+        double r_ii = get_magnitude(Q_i);
+        // save r_ii to matrix R
+        res.R.data[i*res.R.m + i] = r_ii;
+
+        multiply_scalar_vector_inplace(1/r_ii, &Q_i);
+        // Move Q_i back into the corresponding column of the Q matrix 
+        copy_column_to_matrix_inplace(Q_i, &res.Q, i);
+    }
+
+    return res;
+}
+
+// DEBUGGING -----
+
 // Print out a matrix for debugging purposes
 void print_matrix(Matrix X) {
     int i, j;
@@ -118,4 +303,14 @@ void print_matrix(Matrix X) {
         }
         printf("\n");
     }
+}
+
+// Print out a vector for debugging purposes
+void print_vector(Vector x) {
+    int i;
+    printf("PRINTING VECTOR x:\n");
+    for (i = 0; i < x.size; i++) {
+        printf("%lf ", x.data[i]);
+    }
+    printf("\n");
 }
